requirements.txt
python-telegram-bot==20.3
google-generativeai==0.8.3
python-dotenv==1.0.1